Start work on GLUT / GLEW / C11-threads port

+++ port/3ds/types.h
@@ -6,7 +6,8 @@
 
 typedef uint8_t u8;
 typedef uint16_t u16;
-typedef uint32_t u32;
+//typedef uint32_t u32;
+typedef uintptr_t u32; // Fuck it, smea keeps casting back and forth
 typedef uint64_t u64;
 
 typedef int8_t s8;
